Bring visualizer dtos in line with the existing style
