+++ service/src/main/java/de/starwit/service/dto/BoundingBoxDto.java
@@ -1,0 +1,45 @@
+package de.starwit.service.dto;
+
+public class BoundingBoxDto {
+    private double minX;
+    private double minY;
+    private double maxX;
+    private double maxY;
+
+    public double getMinX() {
+        return minX;
+    }
+
+    public void setMinX(double minX) {
+        this.minX = minX;
+    }
+
+    public double getMinY() {
+        return minY;
+    }
+
+    public void setMinY(double minY) {
+        this.minY = minY;
+    }
+
+    public double getMaxX() {
+        return maxX;
+    }
+
+    public void setMaxX(double maxX) {
+        this.maxX = maxX;
+    }
+
+    public double getMaxY() {
+        return maxY;
+    }
+
+    public void setMaxY(double maxY) {
+        this.maxY = maxY;
+    }
+
+    @Override
+    public String toString() {
+        return "BoundingBoxDto [minX=" + minX + ", minY=" + minY + ", maxX=" + maxX + ", maxY=" + maxY + "]";
+    }
+}
