+++ service/src/main/java/de/starwit/service/dto/ShapeDto.java
@@ -1,0 +1,32 @@
+package de.starwit.service.dto;
+
+public class ShapeDto {
+    private int width;
+    private int height;
+
+    public ShapeDto(int width, int height) {
+        this.width = width;
+        this.height = height;
+    }
+
+    public int getWidth() {
+        return width;
+    }
+
+    public void setWidth(int width) {
+        this.width = width;
+    }
+
+    public int getHeight() {
+        return height;
+    }
+
+    public void setHeight(int height) {
+        this.height = height;
+    }
+
+    @Override
+    public String toString() {
+        return "ShapeDto [width=" + width + ", height=" + height + "]";
+    }
+}
